pyproject.toml
# Copyright (c) 2019 Adam Karpierz
# SPDX-License-Identifier: Zlib

[build-system]
build-backend = 'setuptools.build_meta'
requires = ['setuptools>=72.2.0', 'wheel>=0.44.0', 'packaging>=24.1.0', 'tox>=4.18.0']

[project]
name        = 'plasticscm'
version     = '0.3.1a1'
description = 'Python package providing access to the PlasticSCM client API.'
urls.Homepage      = 'https://pypi.org/project/plasticscm/'
urls.Documentation = 'https://plasticscm.readthedocs.io/'
urls.Download      = 'https://pypi.org/project/plasticscm/'
urls.Source        = 'https://github.com/karpierz/plasticscm'
urls.Issues        = 'https://github.com/karpierz/plasticscm/issues'
license = { text = 'zlib/libpng License ; https://opensource.org/license/zlib' }
authors = [
    { name  = 'Adam Karpierz' },
    { email = 'adam@karpierz.net' },
]
maintainers = [
    { name  = 'Adam Karpierz' },
    { email = 'adam@karpierz.net' },
]
keywords = ['PlasticSCM']
classifiers = [
    'Development Status :: 3 - Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: zlib/libpng License',
    'Operating System :: OS Independent',
    'Natural Language :: Polish',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
    'Programming Language :: Python :: 3 :: Only',
    'Programming Language :: Python :: Implementation :: CPython',
    'Programming Language :: Python :: Implementation :: PyPy',
    'Topic :: Software Development :: Libraries :: Python Modules',
]
requires-python = '>=3.8.10,<4.0.0'
dependencies = [
    # mandatory
    'setuptools>=72.2.0',
    'pkg-about>=1.1.8',
    # others
    'atpublic>=5.0.0',
    'requests>=2.32.3',
    'python-dateutil>=2.9.0.post0',
]
dynamic = ['readme']
[project.optional-dependencies]
doc = [
    'Sphinx>=7.1.2',
    'sphinx-toolbox>=3.7.0',
    'sphinx-tabs>=3.4.5',
    'sphinx-copybutton>=0.5.2',
    'sphinxcontrib-spelling>=8.0.0',
    'sphinx-lint>=0.9.1',
    'restructuredtext-lint>=1.4.0',
    'nbsphinx>=0.9.4',
]
test = [
    'deepdiff>=7.0.1',
    'rich>=13.7.1',
    'httmock>=1.4.0',
]

[project.scripts]
#'plasticscm' = 'plasticscm.__main__:main'

[project.gui-scripts]
#'plasticscm-gui' = 'plasticscm:main_gui'

#[project.entry-points.'plasticscm.magical']
#epoint = 'plasticscm:main_epoint'

[tool.setuptools.dynamic]
readme = { file = ['README.rst', 'CHANGES.rst'], content-type = 'text/x-rst; charset=UTF-8' }

[tool.setuptools]
license-files = ['LICENSE']
include-package-data = true
platforms = ['any']
zip-safe = true

[tool.setuptools.packages.find]
namespaces = false
where = ['src']

[tool.setuptools.package-dir]
'' = 'src'
#'plasticscm.tests' = 'tests'

[tool.setuptools.package-data]
plasticscm = [
    'plasticscm.cfg',
]

[tool.setuptools.exclude-package-data]
'*' = ['*.c','*.h','*.cpp','*.hpp','*.cxx','*.hxx','*.pyx','*.pxd']
plasticscm = [
]

[tool.coverage.run]
source = [
    'plasticscm',
    'tests',
]
omit = [
]
data_file = '.tox/coverage/.coverage'
plugins = ['covdefaults']

[tool.coverage.report]
exclude_lines = [
    # Regexes
    # Have to re-enable the standard pragma
    '^\s*if\s+self\.debug\s*:',
    '^\s*if\s+__debug__\s*:',
    '^\s*if\s+(0|False)\s*:',
    '''if\s+__name__.*\s*==\s*['"]__main__['"]\s*:''',
    '^\s*@unittest\.skip\(',
    '^\s*@unittest\.skipIf\(sys\.platform\.startswith\("win"\)',
]
omit = [
    'tests/run.py',
]
skip_covered = false

[tool.coverage.html]
directory = '.tox/coverage/.coverage.html'

[tool.flake8]
filename = ['*.py','*.pyx']
#include = ['tests']
#exclude = ['.tox','*.egg','.git','__pycache__','build','_build','docs/_build','dist']
max-line-length = 99
ignore = ['E126','E203','E221','E251','E302','E701','E702','E731',
          'E122','E127','E128','E222','E272','E241','E266','E226','E704',
          'D100','D101','D102','D103','D104','D400','D401','D202',
          'N806','N802','N803','N801',
          'I100','W503',
          'D419','D106']
# (e.g. 'E4','W') default: 'E121','E123','126','226','E24','704'
#select =
#select = ['E','W','F','N','I']
output-file = '.tox/lint/flake8out.txt'
count = true
#show-pep8,
#show-source
#verbose
#quiet

#
# Configuration(s) for tox
#

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py{38,39,310,311,312,313}, pypy{39,310}, docs
labels =
    py = py{38,39,310,311,312,313}, pypy{39,310}
    prepare = prepare
    coverage = coverage
    lint = lint
    docs = docs
    build  = py{38,39,310,311,312,313}, pypy{39,310}, docs, build
    deploy = py{38,39,310,311,312,313}, pypy{39,310}, docs, build, publish
minversion = 4.18.0
skip_missing_interpreters = true
requires =
    {[base]deps}
    virtualenv>=20.26.3
    tox-backtick>=0.4.6
    tox-tags>=0.2.0

[base]
basepython = python3.12
deps =
    pip>=24.2.0
    setuptools>=72.2.0
    wheel>=0.44.0
packagesubdir = plasticscm

[testenv]
passenv = WINDIR
setenv =
    PYTHONDONTWRITEBYTECODE = 1
commands =
    {envpython} --version
    {envpython} -m tests {posargs}
extras = test
deps =
    {[base]deps}

[testenv:prepare]
basepython = {[base]basepython}
skip_install = true
allowlist_externals =
    cmd
    .build
commands =
    cmd /C if exist .build.cmd .build.cmd

[testenv:coverage]
basepython = {[base]basepython}
commands =
    {envpython} -m coverage erase
    -{envpython} -m coverage run -m tests {posargs}
    {envpython} -m coverage report
    {envpython} -m coverage html
deps =
    {[testenv]deps}
    coverage>=7.6.1
    covdefaults>=2.3.0
    diff-cover>=9.1.1

[testenv:docs]
basepython = {[base]basepython}
commands =
    {envpython} -m sphinxlint --ignore .tox --ignore build --ignore dist
   #{envpython} -m sphinx.apidoc -f {envsitepackagesdir}/{[base]packagesubdir}
    {envpython} -m sphinx.cmd.build -W -a -b html -E   ./docs ./build/docs/html
    {envpython} -m sphinx.cmd.build -W -a -b linkcheck ./docs ./build/docs/html
    {envpython} -m sphinx.cmd.build -W -a -b doctest   ./docs ./build/docs/html
extras = doc
deps =

[testenv:build]
depends = py{38,39,310,311,312,313}, pypy{39,310}, docs
basepython = {[base]basepython}
setenv =
    {[testenv]setenv}
    PKG_PVER=`{envpython} -W ignore -c "import platform ; print(''.join(platform.python_version_tuple()[:2]), end='')" 2> nul`
    PKG_NAME=`{envpython} -W ignore -c "import setuptools ; setuptools._distutils.core._setup_stop_after='config' ; print(setuptools.setup().metadata.get_name(),     end='')" 2> nul`
    PKG_DIST=`{envpython} -W ignore -c "import setuptools ; setuptools._distutils.core._setup_stop_after='config' ; print(setuptools.setup().metadata.get_fullname(), end='')" 2> nul`
commands =
    {envpython} -W ignore -c "from setuptools import setup ; setup()" --quiet sdist --formats=zip
    {envpython} -W ignore -c "from setuptools import setup ; setup()" --quiet bdist_wheel
    # check out for PyPi
    {envpython} -m twine check dist/*
deps =
    {[testenv]deps}
    twine>=5.1.1

[testenv:publish]
depends = build
basepython = {[base]basepython}
skip_install = true
commands =
    # publish on PyPi
    {envpython} -m twine upload dist/*
extras =
deps =
    twine>=5.1.1

[testenv:lint]
basepython = {[base]basepython}
commands =
    {envpython} -m flake8 {envsitepackagesdir}/{[base]packagesubdir}/
extras =
deps =
    {[testenv]deps}
    flake8>=7.1.1
    flake8-pyproject>=1.2.3
    flake8-docstrings>=1.7.0
    pep8-naming>=0.14.1
    flake8-builtins>=2.5.0
    flake8-deprecated>=2.2.1
"""
